test connections url and method
